refactor:Context API 활용하여 리팩토리 진행
