+++ 호담-최규호/2주차/todolist_css/src/App.jsx
@@ -1,43 +1,21 @@
-import { useState } from 'react';
+import { useContext } from 'react';
 import './App.css';
+import { TodoContext } from './context/TodoContext'
 
 function App() {
-  // 투두리스트, 화면에 출력되는 (추가, 삭제, 수정)
-  const [todos, setTodos] = useState([
-    { id: 1, task: '투두 만들어보기'},
-    { id: 2, task: '희연 혜원 혜윤 건 찬민'},
-  ]);
-
-  const [text, setText] = useState('');
-
-  const [editingId, setEditingId] = useState('');
-  const [editText, setEditText] = useState('');
-
-  //랜더링 방지
-  const handleSubmit = (e) => {
-    e.preventDefault();
-  };
-  // 1. 추가하기
-  const addTodo = () => {
-    setTodos((prev) => [
-      ...prev,
-      { id: Math.floor(Math.random() * 100) + 2, task: text}
-    ]);
-    setText('');
-  };
-
-  // 2. 삭제하기
-  const deleteTodo = (id) => {
-    setTodos((prev) => prev.filter((item) => item.id !== id));
-  };
-
-  // 3. 수정하기
-  const updateTodo = (id, text) => {
-    setTodos((prev) => 
-      prev.map((item) => item.id === id ? {...item, task:text}:item)
-    );
-    setEditingId('');
-  };
+  const {
+    todos,
+    text,
+    setText,
+    editingId,
+    setEditingId,
+    editText,
+    setEditText,
+    handleSubmit,
+    addTodo,
+    deleteTodo,
+    updateTodo,
+  } = useContext(TodoContext);
 
   return (
     <>
