+++ 호담-최규호/2주차/todolist_css/src/components/List.jsx
@@ -1,12 +1,11 @@
 const List = (props) => {
     const { tech } = props;
     return (
-        // listStyle을 통해 제거할 수 있습니다.
       <li style={{listStyle: 'none'}}>
-        {tech}
+        { tech }
       </li>
     )
-  }
+}
   
-  export default List
+export default List
