+++ 호담-최규호/2주차/todolist_css/src/main.jsx
@@ -1,10 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
-import './index.css'
+import { TodoContextProvider } from './context/TodoContext.jsx'
 
 createRoot(document.getElementById('root')).render(
-  <StrictMode>
+  <TodoContextProvider>
     <App />
-  </StrictMode>,
+  </TodoContextProvider>,
 )
